리펙토링

+++ hhplus-api/src/main/java/com/hanghe/domain/payment/entity/Payment.java
@@ -6,6 +6,7 @@ import com.hanghe.domain.base.BaseEntity;
 import com.hanghe.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
+import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 
@@ -30,21 +31,21 @@ public class Payment extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private PaymentType type;
 
-    private int amount;
+    private Long amount;
 
-    private Payment(User user, int amount, PaymentType type) {
+    private Payment(User user, Long amount, PaymentType type) {
         validateInputs(user, amount, type);
         this.user = user;
         this.amount = amount;
         this.type = type;
     }
 
-    public static Payment create(User user, int amount, PaymentType type) {
+    public static Payment create(User user, Long amount, PaymentType type) {
         Payment payment = new Payment(user, amount, type);
         return payment;
     }
 
-    private void validateInputs(User user, int amount, PaymentType type) {
+    private void validateInputs(User user, Long amount, PaymentType type) {
         if (user == null) {
             throw new BusinessException(ErrorCode.USER_NOT_FOUND);
         }
