+++ hhplus-api/src/main/java/com/hanghe/domain/payment/service/dto/PaymentUseDTO.java
@@ -2,7 +2,7 @@ package com.hanghe.domain.payment.service.dto;
 
 import com.hanghe.domain.payment.entity.Payment;
 
-public record PaymentUseDTO(Long paymentId, int amount) {
+public record PaymentUseDTO(Long paymentId, Long amount) {
     public static PaymentUseDTO from(Payment payment) {
         return new PaymentUseDTO(
                 payment.getId(),
